pyproject.toml
[project]
name = "sd-imbalance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "hydra-core>=1.3.2",
    "imbalanced-learn>=0.13.0",
    "lightning>=2.5.0.post0",
    "matplotlib>=3.10.0",
    "numpy<2",
    "omegaconf>=2.3.0",
    "pandas>=2.2.3",
    "ruff>=0.9.3",
    "scikit-learn>=1.6.1",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "tqdm>=4.67.1",
    "wandb>=0.19.5",
]
